refactor: refactor custom hooks to their own files
